fix(security): Phase 1 stabilization - SMB hardening, EXIF sanitization, browse allow-list, deps
- SMB: tighten validate_token (reject -, =, , prefixes/separators); replace unsafe libc::geteuid/getegid with rustix safe wrapper; clean up credentials file on mount failure
- EXIF: sanitize_exif_value() blocks @-prefix file-include attack in exiftool overrides; clear error when exiftool binary missing
- Storage browser: refuse /proc, /sys, /dev (raw and post-canonicalize) with audit logging
- AI: replace partial_cmp().unwrap() with NaN-safe Equal fallback in clustering+face sort
- Web deps: bump vite ^5 -> ^7 (resolves esbuild GHSA-67mh-4wv8-2f99 + vite GHSA-4w7w-66w2-5vf9), npm audit fix postcss
- Tests: rename duplicates 38_ai_accuracy/44_motion_photo/52_geolocation_ddt -> 59/60/61
- Clippy: 94 -> 42 warnings via cargo clippy --fix

Refs plan.md Phase 1.1, 1.2, 1.3, 1.4, 1.6, 2.1

+++ server/src/diagnostics/handlers.rs
@@ -134,8 +134,8 @@ pub(crate) fn disk_stats(path: &std::path::Path) -> (u64, u64) {
     unsafe {
         let mut stat: libc::statvfs = std::mem::zeroed();
         if libc::statvfs(c_path.as_ptr(), &mut stat) == 0 {
-            let total = stat.f_blocks as u64 * stat.f_frsize as u64;
-            let available = stat.f_bavail as u64 * stat.f_frsize as u64;
+            let total = stat.f_blocks * stat.f_frsize;
+            let available = stat.f_bavail * stat.f_frsize;
             (total, available)
         } else {
             (0, 0)
